Remove TablePropertiesStore interface
